iterate on IntUiBridgeTextField
